Included using diff images (sizes) for app icons for diff devices and spotlight search. Removed images from root folder as they are copied and included in Images.xcassets folder - Still there may be value in creating and storing these files in a separate images folder. Note storing unused files in Images.xcassetsAppiconset.appiconset folder gives warning about images not assigned.
Fixed conversion warnings from NSInteger to int and use of explicit long conversion while passing NSInteger into (NSString) format strings.

+++ Matchismo/Matchismo/Model/CardGameNotifications.h
@@ -11,6 +11,6 @@
 
 @protocol CardGameNotifications <NSObject>
 
--(void)selectionImpactOfCard:(Card *)card chosen:(BOOL)isChosen otherChosenCards:(NSArray *)otherChosenCards impact:(int)chosenCardsScoreImpact;
+-(void)selectionImpactOfCard:(Card *)card chosen:(BOOL)isChosen otherChosenCards:(NSArray *)otherChosenCards impact:(NSInteger)chosenCardsScoreImpact;
 
 @end
