+++ Matchismo/Matchismo/Model/CardMatchingGame.m
@@ -76,7 +76,7 @@ static const int COST_TO_CHOOSE = 1;
             //calculate impact, update selected score
             card.chosen = YES;
             self.score -= COST_TO_CHOOSE; //for choosing to view a card
-            int cardMatchImpact = [self matchCard:card toOtherCards:prevChosenCards];
+            NSInteger cardMatchImpact = [self matchCard:card toOtherCards:prevChosenCards];
             self.currChosenCardsScore += cardMatchImpact;
 
             if (prevChosenCards.count < (self.matchMode-1)) { // match mode requires more cards to be selected
@@ -124,12 +124,12 @@ static const int COST_TO_CHOOSE = 1;
 {
     int score = 0;
     NSArray * chosenCards = [self currentChosenCards];
-    int chosenCount = chosenCards.count;
+    NSInteger chosenCount = chosenCards.count;
     //match each card against the other cards and sum their matching scores
-    for (int i=chosenCount-1; i>0; i--) {
+    for (NSInteger i=chosenCount-1; i>0; i--) {
         //NSArray * otherCards = [NSArray arrayWithObjects:[(const id[])chosenCards count:i];
         //score += [chosenCards[i] match:otherCards];
-        for (int j=i-1; j>=0; j--) {
+        for (NSInteger j=i-1; j>=0; j--) {
             score += [chosenCards[i] match:@[chosenCards[j]]];
         }
     }
